Alert limiting (#128)
* Add MonitoringAlerts model

* Add monitoring alert router to app.include_router()

* Add GraylogPostRequest model and update create_monitoring_alert function

* Refactor create_monitoring_alert function to initialize MonitoringAlerts with specific fields

* Fix alert_index assignment in create_monitoring_alert function

* Add GraylogPostResponse model and update create_monitoring_alert endpoint

* Add route for running analysis on monitoring alerts

This commit adds a new route `/run_analysis/wazuh` to the `monitoring_alert.py` file. The route is used to run analysis on the monitoring alerts. It retrieves all the monitoring alerts from the database where the customer code matches the provided code and the alert source is WAZUH. The retrieved alerts are then analyzed using the `analyze_wazuh_alerts` function. The commit also includes the addition of the `WazuhAnalysisResponse` model in the `monitoring_alert.py` file. This model represents the response containing the analysis results.

* Refactor monitoring_alert.py and schema/monitoring_alert.py

* Add alert creation and analysis functionality

* Add IrisTags and FilterAlertsRequest models

* Add asset and IOC payload to alert creation in IRIS

* Add alert_status_id field to FilterAlertsRequest and construct_params

* Refactor monitoring_alert module and add new routes and precommit fixes

* Add WazuhIrisAlertContext and WazuhIrisAlertPayload classes and remove duplicate assets from the asset list