Wazuh sca (#216)
* chore: Add endpoints for fetching agent SCA and policy results

* updated dependencies

* updated agent api/types

* copilot-ai-testing

* chore: Update branch name in Docker workflow from network-connectors to main

* feat: Add validation for 'rule_group3' in RuleExcludeRequest

The code changes in `rules.py` add validation for the `rule_group3` field in the `RuleExcludeRequest` class. If `rule_group3` is present, it checks that either `rule_group1` or `rule_group3` is set to "windows". If `rule_group3` is not present, it falls back to the previous validation for `rule_group1`.

This commit message follows the established convention of using a prefix to indicate the type of change (`feat` for a new feature) and provides a clear and concise description of the changes made.

* added sca details

* modal refactor

* added sca item component

* updated VulnerabilityCard

* updated dependencies

* refactor vulnerability

* added sca result list

* added sca result item

* updated sca result item card

* updated sca result details

* updated sca result details

* chore: Update variable name in Wazuh SCA service

* precommit fixes

---------

Co-authored-by: Davide Di Modica <webmaster.ddm@gmail.com>