Process insights (#222)
* feat: Add process_name field to IrisAlertContext model

This commit adds a new optional field, `process_name`, to the `IrisAlertContext` model in the `escalate_alert.py` file. The `process_name` field represents the name of the process associated with the alert. This change improves the alert context by providing additional information about the alert.

* feat: Remove commented code for process analysis in SocFortress route

The code changes in `socfortress.py` remove the commented code for process analysis in the `threat_intel_socfortress` route. This improves the code readability and removes unnecessary code that is not being used.

Note: This commit message follows the established convention of using a prefix to indicate the type of change (`feat` for a new feature) and provides a clear and concise description of the changes made.

* update command to grab windows_firewall active response in MD

* updated dependencies

* added processNameEvaluation api

* refactor soc alerts

* refactor soc alerts

* added Evaluation modal

* updated Evaluation modal

* add process_name to alert creation

* updated Evaluation modal

* updated Evaluation modal

* updated soc alert context tab

* chore: Add endpoint for retrieving artifact recommendation based on alert

* precommit fixes

---------

Co-authored-by: Davide Di Modica <webmaster.ddm@gmail.com>