Duo integration (#247)
* feat: Add DUO integration to SOCFortress

The code changes in `db_populate.py` add the DUO integration to SOCFortress by including the necessary information for DUO authentication keys. This allows users to integrate DUO with SOCFortress for enhanced security.

* feat: Add DUO dashboard to Grafana schema and services

This commit adds the DUO dashboard to the Grafana schema and services. The `dashboards.py` file in the `grafana/schema` directory is updated to include the `DuoDashboard` class with the `DUO_AUTH` dashboard. Similarly, the `dashboards.py` file in the `grafana/services` directory is updated to import and include the `DuoDashboard` in the `provision_dashboards` function. This allows users to provision the DUO dashboard in Grafana for enhanced security monitoring.

* refactor: Rename integration_dup_provision_router to integration_duo_provision_router and add route

* add duo auth logs dashboard template

* refactor: Update DUO integration function names for consistency

This commit updates the function names related to DUO integration for consistency. The function names in the `provision.py` file are changed from `invoke_duo_integration_collection` to `invoke_duo_integration_collect`. Similarly, the function names in the `scheduler.py` file are updated to match the changes. This ensures that all DUO integration functions have consistent naming conventions.

* update: integrations api

* update: CustomerIntegrationActions

* improve: pinned-page component

* precommit fixes

---------

Co-authored-by: Davide Di Modica <webmaster.ddm@gmail.com>