Vuln list all (#252)
* feat: Add endpoint to fetch agent vulnerabilities by severity

The code changes in this commit add a new endpoint to fetch agent vulnerabilities based on their severity. The `get_agent_vulnerabilities` function in the `agents.py` file now accepts a `vulnerability_severity` parameter, allowing users to specify the severity level of the vulnerabilities they want to retrieve. This enhancement improves the flexibility and granularity of vulnerability fetching in the application.

* refactor: Update vulnerability title in process_single_vulnerability function

* refactor: Update index prefix in build_index_set_config function

* feat: Add EDR_SYSTEM_VULNERABILITIES_NEW dashboard to WazuhDashboard enum

* feat: Create vulnerability datasource for Grafana

This code change adds a new function, `create_vulnerability_datasource`, to the `grafana.py` file in the `customer_provisioning/services` directory. This function is used to create a Grafana datasource for Wazuh vulnerabilities using the OpenSearch Data Source. It is specifically designed for Wazuh versions 4.8.0 and above.

* update: dependencies

* update: design token

* improve: PinnedPage component

* update: indecies marquee speed

* update: Vulnerability api

* update: Vulnerabilities tab

* precommit fixes

* feat: Update agent vulnerability fetching logic for Wazuh Manager version 4.8.0 or higher

This code change modifies the `get_agent_vulnerabilities` function in the `agents.py` file. It adds a check for the Wazuh Manager version and if it is 4.8.0 or higher, it fetches the vulnerabilities using the new API. This improves the vulnerability fetching process for agents with newer Wazuh Manager versions.

---------

Co-authored-by: Davide Di Modica <webmaster.ddm@gmail.com>