Velo collect agent org (#256)
* collect velo org per agent testing

* refactor: Collect Velociraptor agents per organization

This code change modifies the `sync_agents_velociraptor` function in the `sync.py` file. It now collects Velociraptor agents per organization, allowing for more efficient synchronization. This refactor improves the organization and management of Velociraptor agents in the system.

* refactor: add velo org to agents table

* add velo org when collecting agents

* refactor: Add velociraptor_org field to BaseBody model

This code change adds the `velociraptor_org` field to the `BaseBody` model in the `artifacts.py` file. The `velociraptor_org` field allows for specifying the organization of the client when collecting artifacts. This enhancement improves the flexibility and accuracy of artifact collection in the system.

* refactor: Extract filename from process_name in SocfortressThreatIntelRequest

This code change adds a validator to the `process_name` field in the `SocfortressThreatIntelRequest` model. The validator extracts the filename from the `process_name` using the `os.path.basename` function. This enhancement improves the consistency and reliability of the filename extraction process in the Socfortress threat intelligence module.

* precommit fixes