Darktrace (#270)
* darktrace auth keys

* darktrace markdown

* markdown cleanup

* provision darktrace initial

* dashboard template

* feat: Add Darktrace dashboard support

The code changes include adding support for the Darktrace dashboard in the Grafana connector. This allows users to provision and use the Darktrace summary dashboard. The changes involve adding the DarktraceDashboard enum and updating the DashboardProvisionRequest model to include the Darktrace dashboard.

Recent user commits:
- dashboard template
- provision darktrace initial
- markdown cleanup
- darktrace markdown
- darktrace auth keys
- precommit fixes

Recent repository commits:
- Refactor artifact recommendation error handling
- Docker compose nuclei (#262)

* feat: Add Darktrace integration to scheduler

* feat: Update Darktrace dashboard title

The code changes include updating the title of the Darktrace dashboard in the Grafana connector. This commit modifies the "title" property in the Darktrace summary.json file to display "DARKTRACE EVENTS AND ALERTS" as the new title.

Note: This commit is related to the recent user commit "feat: Add Darktrace integration to scheduler" and the recent repository commit "feat: Add Darktrace dashboard support".

* feat: Update Darktrace integration service name

The code changes involve updating the integration service name in the `invoke_darktrace.py` file. The previous integration service name "DUO" has been changed to "Darktrace" to accurately reflect the integration.

* update: dependencies

* add: darktrace provision

* precommit fixes

---------

Co-authored-by: Davide Di Modica <webmaster.ddm@gmail.com>