feat: Add processed value for CustomerCodeKeys.CLUSTER_NODE in get_customer_code
This commit adds a processed value for the CustomerCodeKeys.CLUSTER_NODE key in the get_customer_code function of incident_alert.py. The processed value is obtained by calling the get_processed_value method of the CustomerCodeKeys enum. This change ensures that the processed value is returned when the key is CLUSTER_NODE.

Recent user commits:
- Update branch name in Docker workflow
- Add alert title payload to handle_customer_notifications function

Recent repository commits:
- feat: Update branch name in Docker workflow
- feat: Add alert title payload to handle_customer_notifications function